M    psync.py + it sends mail with a summary od the updates. filenames only now. + added options to handle the address to where the mail is send, and the   subject.
+ also, dry-run option. it downloads the new databases, process them, but
  doesn't download any package nor does it updates the databases.
  

M    psyncpkg/core.py
* any extra options passed to __init__ is merged in the dict. ported from
  psyncpkg/drivers/Urpmi.py.
* support for dry-run in getPackage().

M    psyncpkg/drivers/Urpmi.py
M    psyncpkg/drivers/Debian.py
* follow the new Psync.__init__ API.

+++ psyncpkg/core.py
@@ -9,7 +9,8 @@ import os
 from psyncpkg.utils import stat, makedirs, grab, rename
 
 class Psync(object):
-    def __init__ (self, cont=False, consistent=True, limit=20, verbose=False):
+    def __init__ (self, cont=False, consistent=True, limit=20, verbose=False,
+            dryRun=False, **kwargs):
         """_attr means readonly
         """
         self.cont= cont
@@ -18,6 +19,8 @@ class Psync(object):
         self.verbose= verbose
         self.delete= []
         self.failed= []
+        self.dryRun= dryRun
+        self.__dict__.update(kwargs)
 
     def getPackage (self, baseUrl, localDir, fileName, size):
         """ Get one package, making sure it's the right size
@@ -34,7 +37,8 @@ class Psync(object):
         if old:
             if not self.saveSpace:
                 for i in old:
-                    print "%s: wrong version, deleted" % i
+                    if self.verbose:
+                        print "%s: wrong version, deleted" % i
                     unlink (i)
             else:
                 if self.verbose:
@@ -47,12 +51,14 @@ class Psync(object):
         except OSError:
             if self.verbose:
                 print "%s: not here" % _file
-            ans= self.grab (_file, url)
+            if not self.dryRun:
+                ans= self.grab (_file, url)
         else:
             if size is not None and s.st_size!=size:
                 if self.verbose:
                     print "%s: wrong size %d" % (_file, s.st_size)
-                ans= self.grab (_file, url, cont=True)
+                if not self.dryRun:
+                    ans= self.grab (_file, url, cont=True)
             else:
                 if self.verbose:
                     print "%s: already here, skipping" % _file
@@ -78,15 +84,19 @@ class Psync(object):
         # process Packages
         delete= []
         failed= []
+        files= []
         for module in modules:
             # download databases
             for (database, critic) in self.databases (distro, module, arch):
+                # yes: per design, we don't follow the dryRun option here,
+                # but neither the databases will be swaped at the end.
                 found= self.grab (_dir+local+'/'+database,
                            baseurl+'/'+database, cont=self.cont)
 
             # now files
             for filename, size in self.files (_dir+local, distro, module, arch):
                 self.getPackage (baseurl, local, filename, size)
+                files.append (basename(filename))
 
         # summary of failed pkgs
         if failed:
@@ -97,7 +107,7 @@ class Psync(object):
 
         # done
         # clean up/rollover
-        if not self.cont:
+        if not self.cont and not self.dryRun:
             for module in modules:
                 for (database, critic) in self.finalDBs(distro, module, arch):
                     new= local+"/"+database
@@ -110,15 +120,17 @@ class Psync(object):
                     except OSError, e:
                         # better error report!
                         if not critic:
-                            print e, old
+                            print '[Ign]', e, old
                         else:
                             raise e
                 # removedirs (dirname (old))
 
-        if not self.saveSpace:
+        if not self.saveSpace and not self.dryRun:
             # they're old anyways
             for _file in delete:
                 print "unlinking %s" % _file
                 unlink (_file)
+
+        return files
 
     grab= grab
